Validate PANIC
PANIC is a coin created by @panicselling (https://twitter.com/panicselling). 
This coin is created for a vibe and holders of this coin have an 
exclusive opportunity to join the holders’ chat where there are many 
famous names from the Solana ecosystem. In order to join this chat 
you need to have a large amount of $PANIC (you can find out the 
exact number on the official coin page @PANICcoin).