Update PAINT image url in validated-tokens.csv
PAINT has started appearing in the Jupiter Strict list, but the image is not loading. Could it be that the imgur URL we have in the csv is not working for Jupiter? In this commit, I have changed the image URL to use a direct link to the PAINT image file hosted on i.imgur.com, instead of using the imgur.com domain.

Updated image url: https://i.imgur.com/t7CeL6E.png